testing login page/testing login page/requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
pillow==10.1.0
prophet==1.1.5
scikit-learn==1.3.2
flask==2.0.1
pymongo==3.12.0
bcrypt==3.2.0
python-dotenv==0.19.0
flask-login==0.5.0
openpyxl==3.1.2